+++ medical-record/medical-record/src/main/java/com/ivanovp/medical_record/dto/request/DiagnosisRequestDTO.java
@@ -1,0 +1,18 @@
+package com.ivanovp.medical_record.dto.request;
+
+import jakarta.validation.constraints.NotBlank;
+import lombok.AllArgsConstructor;
+import lombok.Data;
+import lombok.NoArgsConstructor;
+
+@Data
+@NoArgsConstructor
+@AllArgsConstructor
+public class DiagnosisRequestDTO {
+
+    @NotBlank(message = "Code is required")
+    private String code;
+
+    @NotBlank(message = "Name is required")
+    private String name;
+}