+++ medical-record/medical-record/src/main/java/com/ivanovp/medical_record/dto/response/DiagnosisResponseDTO.java
@@ -1,0 +1,15 @@
+package com.ivanovp.medical_record.dto.response;
+
+import lombok.AllArgsConstructor;
+import lombok.Data;
+import lombok.NoArgsConstructor;
+
+@Data
+@NoArgsConstructor
+@AllArgsConstructor
+public class DiagnosisResponseDTO {
+
+    private Long id;
+    private String code;
+    private String name;
+}